+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/cohere
 
+## 2.0.9
+
+### Patch Changes
+
+- 0816d3a: feat(provider/cohere): reasoning model support
+
+  Reasoning is now supported for all Cohere models that support it (`command-a-reasoning-08-2025` as of today). See https://docs.cohere.com/docs/reasoning
+
 ## 2.0.8
 
 ### Patch Changes